requirements.txt
anyio==4.4.0
APScheduler==3.6.3
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
cssselect2==0.7.0
dispatcher==1.0
dj-database-url==2.2.0
Django==5.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-telegrambot==1.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
fonttools==4.53.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-decouple==3.8
python-dispatch==0.2.2
python-telegram-bot==13.5
pytz==2024.1
PyYAML==6.0.2
rest-framework-pagination==0.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
